Moved the EscapeDoubleQuotes method to StringExtensions
